refactor: moved AuthInfo interface to common module
