+++ src/__mocks__/modules/auth/services/ClientAuthMock.ts
@@ -1,5 +1,5 @@
 import { ClientAuthDTO } from '@src/modules/auth/interfaces/ClientAuthDTO';
-import { AuthInfo } from '@src/modules/auth/services/interfaces/AuthInfo';
+import { AuthInfo } from '@src/modules/common/interfaces/AuthInfo';
 import { ClientAuthService } from '@src/modules/auth/services/interfaces/ClientAuthService';
 
 export class ClientAuthMock implements ClientAuthService {
