+++ src/__mocks__/modules/auth/services/GenerateTokenMock.ts
@@ -1,4 +1,4 @@
-import { AuthInfo } from '@src/modules/auth/services/interfaces/AuthInfo';
+import { AuthInfo } from '@src/modules/common/interfaces/AuthInfo';
 import { GenerateTokenService } from '@src/modules/auth/services/interfaces/GenerateTokenService';
 
 export class GenerateTokenMock implements GenerateTokenService {
