+++ src/modules/auth/services/interfaces/ClientAuthService.ts
@@ -1,5 +1,5 @@
 import { ClientAuthDTO } from '../../interfaces/ClientAuthDTO';
-import { AuthInfo } from './AuthInfo';
+import { AuthInfo } from '../../../common/interfaces/AuthInfo';
 
 export interface ClientAuthService {
   execute(data: ClientAuthDTO): Promise<AuthInfo>;
