+++ src/modules/auth/services/interfaces/GenerateTokenService.ts
@@ -1,4 +1,4 @@
-import { AuthInfo } from './AuthInfo';
+import { AuthInfo } from '../../../common/interfaces/AuthInfo';
 
 export interface GenerateTokenService {
   execute(authInfo: AuthInfo): Promise<string>;
