+++ src/modules/common/interfaces/AuthInfo.ts
