+++ src/modules/posts/controllers/CreatePostController.ts
@@ -1,4 +1,4 @@
-import { AuthInfo } from '@src/modules/auth/services/interfaces/AuthInfo';
+import { AuthInfo } from '@src/modules/common/interfaces/AuthInfo';
 import { Request, Response } from 'express';
 import { CreatePostService } from '../services/interfaces/CreatePostService';
