+++ src/modules/users/controllers/RemoveUserController.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
 import { RemoveUserService } from '../services/interfaces/RemoveUserService';
 import APIError from '@src/errors/APIError';
-import { AuthInfo } from '@src/modules/auth/services/interfaces/AuthInfo';
+import { AuthInfo } from '@src/modules/common/interfaces/AuthInfo';
 
 export class RemoveUserController implements APIController {
   #removeUserService: RemoveUserService;
