+++ src/modules/users/controllers/UpdateUserController.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
 import { UpdateUserService } from '../services/interfaces/UpdateUserService';
 import APIError from '@src/errors/APIError';
-import { AuthInfo } from '@src/modules/auth/services/interfaces/AuthInfo';
+import { AuthInfo } from '@src/modules/common/interfaces/AuthInfo';
 
 export class UpdateUserController implements APIController {
   #updateUserService: UpdateUserService;
